+++ examples/vue/vue-2/src/components/DynamicallyRenderBuilderPage.vue
@@ -19,7 +19,6 @@
 <script>
 import Vue from 'vue';
 
-import { REGISTERED_COMPONENTS } from '.../init-builder.ts';
 import {
   RenderContent,
   getContent,
@@ -28,6 +27,24 @@ import {
   isEditing,
   isPreviewing,
 } from '@builder.io/sdk-vue/vue2';
+
+// Register your Builder components
+import HelloWorldComponent from './HelloWorld.vue';
+
+export const REGISTERED_COMPONENTS = [
+  {
+    component: HelloWorldComponent,
+    name: 'Hello World',
+    canHaveChildren: true,
+    inputs: [
+      {
+        name: 'text',
+        type: 'string',
+        defaultValue: 'World',
+      },
+    ],
+  },
+];
 
 // TODO: enter your public API key
 const BUILDER_PUBLIC_API_KEY = 'f1a790f8c3204b3b8c5c1795aeac4660'; // ggignore
