--- examples/vue/vue-2/src/components/init-builder.ts
@@ -1,17 +1,0 @@
-// Register your Builder components
-import HelloWorldComponent from './HelloWorld.vue';
-
-export const REGISTERED_COMPONENTS = [
-  {
-    component: HelloWorldComponent,
-    name: 'Hello World',
-    canHaveChildren: true,
-    inputs: [
-      {
-        name: 'text',
-        type: 'string',
-        defaultValue: 'World',
-      },
-    ],
-  },
-];
